Refactoring(server-info): yaml module -> yaml_tools to avoid conflicts
