+++ netutils_linux_hardware/rater.py
@@ -6,7 +6,7 @@ from netutils_linux_hardware.folding import Folding
 from netutils_linux_hardware.memory import Memory
 from netutils_linux_hardware.net import Net
 from netutils_linux_hardware.system import System
-from netutils_linux_hardware.yaml import dict2yaml
+from netutils_linux_hardware.yaml_tools import dict2yaml
 
 
 class Rater(object):
